fixed wrong reconcile for grafana team

+++ api/v1alpha1/grafanaorganization_types.go
@@ -27,7 +27,7 @@ type GrafanaOrganizationSpec struct {
 
 // GrafanaOrganizationStatus defines the observed state of GrafanaOrganization
 type GrafanaOrganizationStatus struct {
-	OrganizationID int `json:"org_id,omitempty"`
+	OrganizationID int64 `json:"org_id,omitempty"`
 }
 
 // +kubebuilder:object:root=true
